requirements.txt
fire==0.5.0
pillow==10.2.0
six==1.15.0
termcolor==2.3.0
